Move dummies to other module.
